fix: change BigDecimal to Double

+++ src/main/java/com/JDBC/Ingredient.java
@@ -54,7 +54,7 @@ public class Ingredient {
         return price;
     }
 
-    public void setPrice(BigDecimal price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
